Move the zipapp main script to a template file
